+++ .pre-commit-config.yaml
@@ -35,10 +35,10 @@ repos:
     hooks:
     - id: black
 
--   repo: https://github.com/pre-commit/mirrors-autopep8
-    rev: v2.0.1
-    hooks:
-    - id: autopep8
+# -   repo: https://github.com/pre-commit/mirrors-autopep8
+#     rev: v2.0.1
+#     hooks:
+#     - id: autopep8
 
 -   repo: https://github.com/PyCQA/flake8
     rev: 6.0.0
