+++ pt_miniscreen/components/enterable_selectable_list.py
@@ -40,7 +40,9 @@ class EnterableSelectableList(SelectableList, Navigable, Enterable, HasGutterIco
         return None
 
     def bottom_gutter_icon(self):
-        if isinstance(self.selected_row.enterable_component, Actionable):
+        if isinstance(self.selected_row, Enterable) and isinstance(
+            self.selected_row.enterable_component, Actionable
+        ):
             return get_image_file_path("gutter/tick.png")
 
         if self.can_enter:
